Sync bootstrap code

+++ src/bootstrap/handler.ts
@@ -24,10 +24,14 @@ import {
 } from "./headers.ts";
 import { parseRequestInvocationMetadata } from "./invocation_metadata.ts";
 import { RequestMetrics } from "./metrics.ts";
-import { requestStore } from "./request_store.ts";
+import {
+  getResponseWithRequestStoreCleanup,
+  requestStore,
+} from "./request_store.ts";
 import { Router } from "./router.ts";
 import type { Functions } from "./stage_2.ts";
 import { ErrorType, PassthroughError, UserError } from "./util/errors.ts";
+import "./globals/types.ts";
 
 interface HandleRequestOptions {
   fetchRewrites?: Map<string, string>;
@@ -50,11 +54,20 @@ export const handleRequest = async (
   const environment = getEnvironment();
   const logger = systemLogger.withRequestID(id);
 
-  // we already parse this a bit later. doing it here is a tiny bit expensive,
+  // We already parse this a bit later. Doing it here is a tiny bit expensive,
   // please remove this once you don't need it anymore.
   const featureFlags = parseFeatureFlagsHeader(
     req.headers.get(InternalHeaders.FeatureFlags),
   );
+
+  // In the first iteration of the request context tracking logic, we deleted
+  // from the request store as soon as a response was returned. This caused
+  // issues with invocations that needed access to the context after the
+  // response had been returned (e.g. emitting a log line in a streaming
+  // function). This new mechanism lets us be a bit smarter about when we
+  // delete the request by looking at when the response body is flushed.
+  const useImprovedRequestStoreCleanup =
+    featureFlags[FeatureFlag.ImprovedRequestStoreCleanup];
 
   // A collector of all the functions invoked by this chain or any sub-chains
   // that it triggers.
@@ -192,7 +205,7 @@ export const handleRequest = async (
         .debug("Edge function returned cacheable cache-control headers");
     }
 
-    return mutateHeaders(response, (headers) => {
+    const responseWithMutatedHeaders = mutateHeaders(response, (headers) => {
       // An issue with `Deno.serve` body compression is causing browsers to
       // buffer responses that should be streamed. As a temporary workaround,
       // we ensure that the response has `cache-control: no-transform`.
@@ -207,6 +220,12 @@ export const handleRequest = async (
 
       metrics.writeHeaders(headers);
     });
+
+    if (useImprovedRequestStoreCleanup) {
+      return getResponseWithRequestStoreCleanup(responseWithMutatedHeaders, id);
+    }
+
+    return responseWithMutatedHeaders;
   } catch (error) {
     let errorString = String(error);
 
@@ -249,6 +268,10 @@ export const handleRequest = async (
         .log("uncaught exception while handling request");
     }
 
+    if (useImprovedRequestStoreCleanup && id) {
+      requestStore.delete(id);
+    }
+
     const response = new Response(errorString, {
       status: 500,
       headers: {
@@ -260,7 +283,7 @@ export const handleRequest = async (
 
     return response;
   } finally {
-    if (id) {
+    if (!useImprovedRequestStoreCleanup && id) {
       requestStore.delete(id);
     }
   }
